requirements.txt
pyautogen
ollama
itakello-logging
pymongo
inquirer3
prompt_toolkit
nltk
inflect
pytest
pytest-mock
python-dotenv
asynci